author check added

+++ app/api/category/[id]/route.js
@@ -29,21 +29,32 @@ export const PUT = async (req, { params }) => {
     const { title } = await req.json()
     await connectToDB()
 
+    const category = await Category.findOne({ _id: params.id })
+
+    if (!category) {
+      return new Response(
+        JSON.stringify({ message: 'There is no category with this id' }),
+        { status: 400 }
+      )
+    }
+    if (session.user.id != category.author) {
+      return new Response(
+        JSON.stringify({
+          message: 'You are not allowed to update this category',
+        }),
+        { status: 400 }
+      )
+    }
+
     const updatedCategory = await Category.findOneAndUpdate(
       { _id: params.id },
       { title },
       { new: true }
     )
 
-    if (!updatedCategory) {
-      return new Response(
-        { message: 'There is no category with this id' },
-        { status: 400 }
-      )
-    }
-
     return new Response(JSON.stringify(updatedCategory), { status: 200 })
   } catch (error) {
+    console.log(error)
     return new Response(error.message, { status: 500 })
   }
 }
@@ -59,6 +70,23 @@ export const DELETE = async (req, { params }) => {
 
   try {
     await connectToDB()
+
+    const category = await Category.findOne({ _id: params.id })
+
+    if (!category) {
+      return new Response(
+        JSON.stringify({ message: 'There is no category with this id' }),
+        { status: 400 }
+      )
+    }
+    if (session.user.id != category.author) {
+      return new Response(
+        JSON.stringify({
+          message: 'You are not allowed to delete this category',
+        }),
+        { status: 400 }
+      )
+    }
 
     await Category.deleteOne({ _id: params.id })
